extracted out repository picker base component
